moving default_sample_fields
